% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_primary_key.R
\name{add_primary_key}
\alias{add_primary_key}
\title{Adds primary key to dataframe in position 1}
\usage{
add_primary_key(dataframe, pkey_column_name)
}
\arguments{
\item{pkey_column_name}{name of pkey column}
}
\description{
Adds primary key to dataframe in position 1
}
