% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_col.R
\name{aggregate_col}
\alias{aggregate_col}
\title{Aggregate column into a single string}
\usage{
aggregate_col(data, col, collapse)
}
\arguments{
\item{data}{dataframe}

\item{col}{unquoted column to aggregate}

\item{collapse}{collapse argument for the paste function}
}
\description{
Aggregate column into a single string
}
