% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-na.R
\name{all_is_na}
\alias{all_is_na}
\title{FUNCTION_TITLE}
\usage{
all_is_na(vector)
}
\arguments{
\item{vector}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
