% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-classes.R
\name{all_numeric_cols}
\alias{all_numeric_cols}
\title{FUNCTION_TITLE}
\usage{
all_numeric_cols(data)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
