% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-na.R
\name{all_same_value}
\alias{all_same_value}
\title{FUNCTION_TITLE}
\usage{
all_same_value(vector, na.rm = FALSE)
}
\arguments{
\item{vector}{PARAM_DESCRIPTION}

\item{na.rm}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
