% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_dataframe_if_new_obs.R
\name{append_dataframe_if_new_obs}
\alias{append_dataframe_if_new_obs}
\title{Adds an observation with tinmestamp to a column by given id where column names match environment variables}
\usage{
append_dataframe_if_new_obs(dataframe, id_variable, new_obs_variable,
  timestamp_col_name)
}
\arguments{
\item{id_variable}{name of variable and object name that designates a character vector of ids}

\item{new_obs_variable}{name of variable and object name that provides the input}

\item{timestamp_col_name}{name of the timestamp column that will be timestampped if the observation needs to be added}
}
\description{
Adds an observation with tinmestamp to a column by given id where column names match environment variables
}
