% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R
\name{arrange_by_nchar}
\alias{arrange_by_nchar}
\title{Arrange a Column by Character Number}
\usage{
arrange_by_nchar(data, col, desc = FALSE)
}
\arguments{
\item{data}{A dataframe or tibble.}

\item{col}{Target column.}

\item{desc}{Return sorted in descending order?}

\item{.by_group}{If TRUE, will sort first by grouping variable. Applies to grouped data frames only.}
}
\description{
Order a dataframe based on the number of characters in the column provided as an argument.
}
\seealso{
\code{\link[dplyr]{arrange_all}}
}
