% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R
\name{arrange_num}
\alias{arrange_num}
\title{Arrange by a Given Column as Numeric}
\usage{
arrange_num(data, col, desc = FALSE, .by_group = FALSE)
}
\arguments{
\item{data}{A dataframe or tibble.}

\item{col}{Target column.}

\item{desc}{Return sorted in descending order?}

\item{.by_group}{If TRUE, will sort first by grouping variable. Applies to grouped data frames only.}
}
\description{
Arrange by a Given Column as Numeric
}
\examples{
library(tidyverse)
data("mtcars")

# As Numeric
arrange_num(mtcars, col = carb) 
arrange_num(mtcars, col = carb, desc = TRUE)

# As Integer
arrange_int(mtcars, col = carb) 

# As Double
arrange_dbl(mtcars, col = carb) 


# Native class is maintained
mtcars \%>\%
        mutate_all(as.character) \%>\%
        arrange_num(carb)

mtcars \%>\%
        mutate_all(as.character) \%>\%
        arrange_num(carb) \%>\%
        select(carb) \%>\%
        unlist() \%>\%
        class()
}
\seealso{
\code{\link[dplyr]{arrange_at}},\code{\link[dplyr]{vars}}

Other arrange functions: 
\code{\link{arrange_dbl}()},
\code{\link{arrange_int}()}
}
\concept{arrange functions}
