% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_mr_clean.R
\name{call_mr_clean}
\alias{call_mr_clean}
\title{Mutates all columns of a dataframe to character and trims all white spaces}
\usage{
call_mr_clean(dataframe)
}
\description{
Mutates all columns of a dataframe to character and trims all white spaces
}
