% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-classes.R
\name{class_cols_ff}
\alias{class_cols_ff}
\title{FUNCTION_TITLE}
\usage{
class_cols_ff(class)
}
\arguments{
\item{class}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[purrr]{keep}}
}
