% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce_.R
\name{coalesce_at}
\alias{coalesce_at}
\title{Coalesce 1+ Fields to an Existing Field}
\usage{
coalesce_at(data, at_var, ..., remove = TRUE)
}
\arguments{
\item{at_var}{character vector of length 1 of the column that the coalesce function will be combine to}
}
\description{
This is 2 of 2 coalesce functions.
}
\seealso{
\code{\link[dplyr]{tidyeval-compat}},\code{\link[dplyr]{mutate_all}},\code{\link[dplyr]{select}},\code{\link[dplyr]{mutate}}
}
