% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce_into.R
\name{coalesce_into}
\alias{coalesce_into}
\title{Coalesce 1+ Fields to a New Field}
\usage{
coalesce_into(.data, into, ..., remove = FALSE)
}
\description{
This is 1 of 2 coalesce functions. This one takes fields that are not the same as the proposed new column name. For example, a dataframe with fields "A", "B", and "C" can be coalesced into a field "D". If all the values need to be coalesced into an existing field such as "A", "B", and "C" to "A", use coalesce_at function instead.
}
