% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-classes.R
\name{col_class}
\alias{col_class}
\title{FUNCTION_TITLE}
\usage{
col_class(data, col)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{col}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[dplyr]{select}}
}
