% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-classes.R
\name{col_classes}
\alias{col_classes}
\title{FUNCTION_TITLE}
\usage{
col_classes(data)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
