% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_to_list.R
\name{cols_to_list}
\alias{cols_to_list}
\title{Select cols and convert them to a list}
\usage{
cols_to_list(dataframe, ..., id_column_name = NULL, named = TRUE)
}
\arguments{
\item{id_column_name}{optional string of the column designated as the identifier that should remain in the output.}

\item{named}{the returned list is named with the column name}
}
\description{
This function takes a dataframe and breaks it down into a list by column. The columns can be explicitly provided or if ... is empty, the entire dataframe will be broken down.
}
