% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataframe_skeleton.R
\name{create_dataframe_skeleton}
\alias{create_dataframe_skeleton}
\title{Takes a character vector and creates a 0-row dataframe with the vectors as column names}
\usage{
create_dataframe_skeleton(column_names)
}
\description{
Takes a character vector and creates a 0-row dataframe with the vectors as column names
}
