% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_new_id.R
\name{create_new_pk}
\alias{create_new_pk}
\title{Creates the next primary key}
\usage{
create_new_pk(dataframe, id_variable)
}
\arguments{
\item{id_variable}{column that contains the primary key}
}
\description{
Creates the next primary key
}
