% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{date_to_col}
\alias{date_to_col}
\title{Add a Date Column to Position 1}
\usage{
date_to_col(data, var = "date")
}
\arguments{
\item{data}{A dataframe or tibble.}

\item{var}{Character string of new column name.}
}
\description{
Add a Date Column to Position 1
}
\examples{
library(tidyverse)
test_data <- 
        tibble(Fruits = 
                       c("Apples and pears",
                         "Citrus – oranges, grapefruits, mandarins and limes",
                         "Stone fruit – nectarines, apricots, peaches and plums",
                         "Tropical and exotic – bananas and mangoes",
                         "Berries – strawberries, raspberries, blueberries, kiwifruit and passionfruit")
        )

# Datetime 
datetime_to_col(data = test_data,
                var = "test_datetime")


# Date 
date_to_col(data = test_data,
            var = "test_date")
}
\seealso{
\code{\link[dplyr]{mutate}},\code{\link[dplyr]{select}},\code{\link[dplyr]{reexports}}
}
