% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated1.R
\name{dated}
\alias{dated}
\title{Add the current date}
\usage{
dated(punct = FALSE)
}
\arguments{
\item{punct}{default is FALSE. If true, returns date in YYYYmmdd format. Otherwise YYYY-mm-dd format.}
}
\description{
Add the current date
}
