% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_.R
\name{deselect_if_all_na}
\alias{deselect_if_all_na}
\title{Deselect columns that are all NA}
\usage{
deselect_if_all_na(.data)
}
\description{
Deselect columns that are all NA
}
\seealso{
\code{\link[dplyr]{select_all}}
}
