% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_.R
\name{deselect_if_all_same}
\alias{deselect_if_all_same}
\title{Deselect columns that are all NA}
\usage{
deselect_if_all_same(data, na.rm = FALSE)
}
\description{
Deselect columns that are all NA
}
\examples{
library(tidyverse)

test_data <-
        tibble(A = rep(NA_character_, 5),
               B = LETTERS[1:5],
               C = 1:5,
               D = rep("Z", 5))

test_data
deselect_if_all_na(test_data)
deselect_if_all_same(test_data)
}
\seealso{
\code{\link[dplyr]{select_all}}
}
