% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_all_grepl_any.R
\name{filter_all_grepl_any}
\alias{filter_all_grepl_any}
\title{Filter all the variables for any matches to the grepl phrase}
\usage{
filter_all_grepl_any(
  dataframe,
  grepl_phrase,
  evaluates_to = TRUE,
  ignore.case = TRUE
)
}
\arguments{
\item{dataframe}{input dataframe}

\item{grepl_phrase}{phrase that is being filtered for}

\item{evaluates_to}{whether the filter is for a grepl evaluation of TRUE or FALSE. Defaults to TRUE.}

\item{ignore.case}{boolean the ignore.case argument of grepl function}
}
\description{
This function will return any TRUE findings of the indicated grepl phrase
}
\seealso{
\code{\link[dplyr]{filter_all}}
}
