% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_at_grepl.R
\name{filter_at_grepl}
\alias{filter_at_grepl}
\title{Filter at a variable that contains a phrase}
\usage{
filter_at_grepl(dataframe, col, grepl_phrase)
}
\description{
Filter at a variable that contains a phrase
}
