% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_at_grepl_any.R
\name{filter_at_grepl_any}
\alias{filter_at_grepl_any}
\title{Filter at a variable that contains or doesn't contain a phrase}
\usage{
filter_at_grepl_any(
  dataframe,
  ...,
  grepl_phrase,
  evaluates_to = TRUE,
  ignore.case = TRUE
)
}
\arguments{
\item{dataframe}{input dataframe}

\item{grepl_phrase}{phrase that is being filtered for}

\item{evaluates_to}{whether the filter is for a grepl evaluation of TRUE or FALSE. Defaults to TRUE.}

\item{ignore.case}{boolean the ignore.case argument of grepl function}

\item{col}{column to filter}
}
\description{
Filter at a variable that contains or doesn't contain a phrase
}
