% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_words_in_str.R,
%   R/filter_by_words_in_str_brakes_before_zero.R
\name{filter_by_words_in_str}
\alias{filter_by_words_in_str}
\title{Loop a filter over words in a string
This function splits a string based on the provide split argument, and recursively filters a dataframe for those words at a given column.}
\usage{
filter_by_words_in_str(string, split, dataframe, col)

filter_by_words_in_str(string, split, dataframe, col)
}
\description{
Loop a filter over words in a string
This function splits a string based on the provide split argument, and recursively filters a dataframe for those words at a given column.

Loop a filter over words in a string until there are zero rows
This function splits a string based on the provide split argument, and recursively filters a dataframe for those words at a given column. It differs from the filter_by_words_in_str function because it returns a dataframe if the loop ends up evaluating to zero rows.
}
