% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated2.R, R/filter_first_row.R
\name{filter_first_row}
\alias{filter_first_row}
\title{Filter for the first row}
\usage{
filter_first_row(.data, invert = FALSE)

filter_first_row(.data, invert = FALSE)
}
\description{
Filter for the first row

Filter for the first row
}
\keyword{internal}
