% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_for.R
\name{filter_for}
\alias{filter_for}
\title{Filter a column for a set of values}
\usage{
filter_for(dataframe, filter_col, inclusion_vector, invert = FALSE)
}
\arguments{
\item{filter_col}{column targeted for filtering}

\item{inclusion_vector}{a vector of values to be included}
}
\description{
This is a shortcut to using the dplyr::filter({col} %in% {vector}) function along with its inverse.
}
