% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_for_all}
\alias{filter_for_all}
\title{Filter a column for a set of values}
\usage{
filter_for_all(data, ..., vector, invert = FALSE)
}
\arguments{
\item{vector}{a vector of values to be included}

\item{cols}{column targeted for filtering}
}
\description{
This is a shortcut to using the dplyr::filter({col} %in% {vector}) function along with its inverse.
}
