% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{filter_for_date}
\alias{filter_for_date}
\title{Filters a dataframe for a date in YYYY-mm-dd format in a timestamp column}
\usage{
filter_for_date(.data, timestamp_col, date)
}
\arguments{
\item{.data}{.data with timestamp column}

\item{timestamp_col}{column where timestamp has date and/or datetime data that is in the format YYYY-mm-dd}

\item{date}{date as character string of length one in format YYYY-mm-dd}
}
\value{
dataframe filtered for the date
}
\description{
Filters a dataframe for a date in YYYY-mm-dd format in a timestamp column
}
\seealso{
\code{\link[dplyr]{filter_all}},\code{\link[dplyr]{all_vars}}
}
