% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_for_vector.R
\name{filter_for_vector}
\alias{filter_for_vector}
\title{Filter for observations using a character vector of included values}
\usage{
filter_for_vector(dataframe, filter_col, inclusion_vector)
}
\arguments{
\item{filter_col}{column targetted for filtering}

\item{inclusion_vector}{character vector of values to be included}
}
\description{
Filter for observations using a character vector of included values
}
