% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_for_vector_function.R
\name{filter_in_vector}
\alias{filter_in_vector}
\title{Filter for observations using a character vector of included values}
\usage{
filter_in_vector(dataframe, filter_col, inclusion_vector)
}
\arguments{
\item{filter_col}{column targetted for filtering}

\item{inclusion_vector}{character vector of values to be included}
}
\description{
Filter for observations using a character vector of included values
}
