% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_for_current_records.R
\name{filter_most_recent_obs_by_group}
\alias{filter_most_recent_obs_by_group}
\title{Filters most recent values in dataframe based on timestamp or date data}
\usage{
filter_most_recent_obs_by_group(dataframe, ..., index_time_col,
  time_format = c("date", "timestamp"))
}
\arguments{
\item{...}{grouping variables}

\item{index_time_col}{variable that will be used to filter by time or datettime}
}
\description{
Filters most recent values in dataframe based on timestamp or date data
}
