% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_out_row_number.R
\name{filter_out_row_number}
\alias{filter_out_row_number}
\title{Removes a row by its number}
\usage{
filter_out_row_number(dataframe, row.number)
}
\arguments{
\item{row.number}{integer}
}
\description{
Removes a row by its number
}
