% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_out_vector_function.R
\name{filter_out_vector}
\alias{filter_out_vector}
\title{Filter out observations using a character vector of excluded values}
\usage{
filter_out_vector(dataframe, filter_col, exclusion_vector)
}
\arguments{
\item{filter_col}{column targetted for filtering}

\item{exclusion_vector}{character vector of values to be excluded}
}
\description{
Filter out observations using a character vector of excluded values
}
