% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterByList.R
\name{filterByList}
\alias{filterByList}
\title{Filter By a List and Other Filters}
\usage{
filterByList(.data, ...)
}
\description{
A List object with name as filter fields and values can be used by splicing it with !!! when included as an argument.
}
