% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterList.R
\name{filterList}
\alias{filterList}
\title{Filter by a list or vector}
\usage{
filterList(.data, ...)
}
\arguments{
\item{.data}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
                     
if(interactive()){
testDf <- data.frame(A = 1:3,
                     B = 4:6)
                     
print(testDf)

applyFilters <- list("A == 2",
                     "B == 5") 
                     
testDf \%>\% 
    filterList(applyFilters)
 }
}
\seealso{
\code{\link[rlang]{list2}},\code{\link[rlang]{parse_expr}}
}
