% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_names}
\alias{format_names}
\title{Standardize List Names}
\usage{
format_names(
  list,
  order = c("btrim", "tolower", "uscore", "rm_trailing_punct", "rm_leading_punct")
)
}
\description{
This function takes a dataframe and changes column names in the following order: 1) Trims whitespace on both sides, 2) converts to uppercase, 3) replaces punctuation of any length with a single underscore, 4) removes trailing underscores from native column names that often times have trailing punctuation (usually decimals)
}
\seealso{
\code{\link[purrr]{map}}
 \code{\link[centipede]{trimws}}
 \code{\link[stringr]{str_replace}},\code{\link[stringr]{str_remove}}
}
