% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na-utils.R
\name{grep_na_str}
\alias{grep_na_str}
\title{FUNCTION_TITLE}
\usage{
grep_na_str(vector)
}
\arguments{
\item{vector}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
