% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by_aggregate.R
\name{group_by_aggregate}
\alias{group_by_aggregate}
\title{Aggregate Function}
\usage{
group_by_aggregate(dataframe, ..., agg.col, collapse = "|")
}
\description{
Aggregate Function
}
\seealso{
\code{\link[dplyr]{tidyeval-compat}},\code{\link[dplyr]{group_by}},\code{\link[dplyr]{summarise_all}}
}
