% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by_unique_aggregate.R
\name{group_by_unique_aggregate}
\alias{group_by_unique_aggregate}
\title{Aggregate Function}
\usage{
group_by_unique_aggregate(dataframe, ..., agg.col, collapse = "|")
}
\description{
Aggregate Function
}
