% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped.R
\name{grouped_paste}
\alias{grouped_paste}
\title{Aggregate Function}
\usage{
grouped_paste(data, ..., unique = TRUE, paste_col, collapse = "|")
}
\description{
Aggregate Function
}
\seealso{
\code{\link[dplyr]{tidyeval-compat}},\code{\link[dplyr]{group_by}},\code{\link[dplyr]{summarise_all}}
}
