% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_pkey.R
\name{impute_pkey}
\alias{impute_pkey}
\title{Imputes true NA values in given columns with previous value (starting at position 2)}
\usage{
impute_pkey(dataframe, pkey_column)
}
\arguments{
\item{...}{columns to be imputed}
}
\description{
Imputes true NA values in given columns with previous value (starting at position 2)
}
