% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_with_carry_forward.R
\name{impute_with_carry_forward}
\alias{impute_with_carry_forward}
\title{Imputes true NA values in given columns with previous value (starting at position 2)}
\usage{
impute_with_carry_forward(dataframe, ...)
}
\arguments{
\item{...}{columns to be imputed}
}
\description{
Imputes true NA values in given columns with previous value (starting at position 2)
}
