% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-classes.R
\name{int_cols}
\alias{int_cols}
\title{FUNCTION_TITLE}
\usage{
int_cols(data)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[purrr]{keep}}
}
