% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_join_with_list_names.R
\name{left_join_with_list_names}
\alias{left_join_with_list_names}
\title{Left joins dataframes in the same named list}
\usage{
left_join_with_list_names(list, x_name, y_name, by = NULL,
  suffix = c(".x", ".y"), ...)
}
\arguments{
\item{list}{list of dataframes of length of 2 or greater}

\item{x_name}{name of the dataframe x}

\item{y_name}{name of dataframe y}
}
\description{
Left joins dataframes in the same named list
}
