% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_list.R
\name{list_to_tibble}
\alias{list_to_tibble}
\title{List to Tibble}
\usage{
list_to_tibble(list)
}
\description{
Convert a list of vectors into a dataframe.
}
\seealso{
Other dataframe <> list functions: 
\code{\link{cols_to_list}()}
}
\concept{dataframe <> list functions}
