% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_summary_fun_library.R
\name{load_sum_fun_library}
\alias{load_sum_fun_library}
\title{Load Summary Function Library}
\usage{
load_sum_fun_library()
}
\description{
Load Summary Function Library
}
\seealso{
\code{\link[centipede]{no_na}}
}
