% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated2.R
\name{loadSummaryFnLibrary}
\alias{loadSummaryFnLibrary}
\title{Load Summary Function Library}
\usage{
loadSummaryFnLibrary()
}
\description{
Load Summary Function Library
}
\seealso{
\code{\link[centipede]{no_na}}
}
\keyword{internal}
