% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_column_prefixes.R
\name{mask_column_prefixes}
\alias{mask_column_prefixes}
\title{Removes lengthy column prefixes that match the R object name for better visualization}
\usage{
mask_column_prefixes(dataframe)
}
\description{
Removes lengthy column prefixes that match the R object name for better visualization
}
