% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated2.R
\name{mutate_all_as_char}
\alias{mutate_all_as_char}
\title{General Dataframe Cleanup
Convert all columns to character class and trim all left and right white spaces.}
\usage{
mutate_all_as_char(dataframe)
}
\arguments{
\item{dataframe}{input dataframe}
}
\description{
General Dataframe Cleanup
Convert all columns to character class and trim all left and right white spaces.
}
\keyword{internal}
