% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated3.R
\name{mutate_all_na_to_blank}
\alias{mutate_all_na_to_blank}
\title{Substitute all true NA values in a dataframe as blank}
\usage{
mutate_all_na_to_blank(dataframe, include_na_as_string = TRUE)
}
\description{
Substitute all true NA values in a dataframe as blank
}
