% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_all_remove_multibyte_chars.R
\name{mutate_all_rm_multibyte_chars}
\alias{mutate_all_rm_multibyte_chars}
\title{Remove all multibyte characters in the dataframe}
\usage{
mutate_all_rm_multibyte_chars(dataframe)
}
\description{
Remove all multibyte characters in the dataframe
}
