% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_all_.R
\name{mutate_all_trimws}
\alias{mutate_all_trimws}
\title{Trim Whitespace of all Character Columns}
\usage{
mutate_all_trimws(
  data,
  which = c("both", "left", "right"),
  whitespace = "[ \\t\\r\\n]"
)
}
\arguments{
\item{data}{A dataframe or tibble.}

\item{which}{a character string specifying whether to remove both
    leading and trailing whitespace (default), or only leading
    (\code{"left"}) or trailing (\code{"right"}).  Can be abbreviated.}

\item{whitespace}{a string specifying a regular expression to match
    (one character of) \dQuote{white space}, see Details for
    alternatives to the default.}
}
\description{
Trim Whitespace of all Character Columns
}
\examples{
library(tidyverse)
test_data <- 
        tibble(A = sample(c(NA_integer_, 1:3), size = 10, replace = TRUE),
               B = sample(c(NA_integer_, 4:6), size = 10, replace = TRUE),
               C = sample(c(" apple ", "banana ", "    orange", NA_character_), size = 10, replace = TRUE))

# Mutate all to character
mutate_all_char(test_data)


# Trim whitespace at all character cols 
mutate_all_trimws(test_data,
                  which = "both")

mutate_all_trimws(test_data,
                  which = "left")

mutate_all_trimws(test_data,
                  which = "right")
}
\seealso{
\code{\link[dplyr]{mutate_at}}
}
