% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_date_cols_to_datetype.R
\name{mutate_date_cols_to_datetype}
\alias{mutate_date_cols_to_datetype}
\title{Mutates date columns to date type}
\usage{
mutate_date_cols_to_datetype(dataframe, ...)
}
\description{
Mutates date columns to date type
}
