% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated2.R
\name{mutate_if_not_exist}
\alias{mutate_if_not_exist}
\title{Adds a column if it doesn't exist in the dataframe}
\usage{
mutate_if_not_exist(dataframe, column_name, value = "")
}
\arguments{
\item{column_name}{character string of new column name}

\item{value}{character vector of length 1 of the new value. Defaults to "".}
}
\description{
Adds a column if it doesn't exist in the dataframe
}
\keyword{internal}
