% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_new_cols_if_not_exist_by_robj_name.R
\name{mutate_new_cols_if_not_exist_by_robj_name}
\alias{mutate_new_cols_if_not_exist_by_robj_name}
\title{Adds a column if it doesn't exist in the dataframe for each object in the global environment}
\usage{
mutate_new_cols_if_not_exist_by_robj_name(dataframe, robj_pattern)
}
\arguments{
\item{robj_pattern}{regular expression pattern of robj names to be added to the dataframe}
}
\description{
Adds a column if it doesn't exist in the dataframe for each object in the global environment
}
