% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{mutate_timestamp_column}
\alias{mutate_timestamp_column}
\title{Add timestamp to the entire dataframe at position 1 with the provided column name}
\usage{
mutate_timestamp_column(dataframe, new_col_name = "TIMESTAMP")
}
\arguments{
\item{new_col_name}{defaults to "TIMESTAMP"}
}
\description{
Add timestamp to the entire dataframe at position 1 with the provided column name
}
