% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_all_to_na.R
\name{normalize_all_to_na}
\alias{normalize_all_to_na}
\title{Normalize a set of values to a true NA}
\usage{
normalize_all_to_na(.data, blanks = TRUE)
}
\description{
This functions takes all the fields in a dataframe and replaces the "NA" string with NA_character_. If blanks is set to TRUE, blanks of character length of 0 are also replaced with NA_character_.
}
