% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_.R
\name{observation_count}
\alias{observation_count}
\title{Get the number of times a row appears in a dataframe}
\usage{
observation_count(data, desc = TRUE)
}
\arguments{
\item{data}{A dataframe or tibble.}

\item{desc}{If TRUE, the output is arranged in descending order. Otherwise it is arranged in ascending order.}
}
\value{
Ungrouped dataframe with all input columns with the addition of an `n` column for the count.
}
\description{
Not to be confused with a total row count of a dataframe (ie `nrow()`), this is a shortcut for `group_by_all()` followed by `count()`.
}
\examples{
library(tidyverse)
test_data <- 
        tibble(
                Group = sample(c("Apple", "Pear"), size = 10, replace = TRUE),
                A     = sample(c(NA_integer_, 1:3), size = 10, replace = TRUE),
                B     = sample(c(NA_integer_, 4:6), size = 10, replace = TRUE),
                C     = sample(c(NA_real_, seq(from = 6.01, to = 6.09, by = 0.01)), size = 10, replace = TRUE)
        )


summarize_variables(data = test_data,
                    incl_num_calc = FALSE)

summarize_variables(data = test_data,
                    incl_num_calc = TRUE)

summarize_variables(data = test_data,
                    incl_num_calc = TRUE,
                    grouper = Group)

summarize_variables(data = test_data,
                    incl_num_calc = FALSE,
                    grouper = Group)


observation_count(data = test_data)


value_count(data = test_data)
value_count(data = test_data,
            grouper = Group)
}
\seealso{
\code{\link[dplyr]{group_by_all}},\code{\link[dplyr]{count}},\code{\link[dplyr]{group_by}},\code{\link[dplyr]{arrange}}

Other summary functions: 
\code{\link{summarize_variables}()}
}
\concept{summary functions}
