% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_parsing.R
\name{parse_date_col}
\alias{parse_date_col}
\title{Parse a Date Column}
\usage{
parse_date_col(
  data,
  col,
  quiet = TRUE,
  tz = NULL,
  locale = Sys.getlocale("LC_TIME")
)
}
\description{
Parse a Date Column
}
\details{
Includes evaluation of the variables as dates with origins of 1900-01-01 and 1970-01-01. Since the evaluation of origin is temperamental (ie 91884 is parsed into a date in 1997 which would be difficult to weed out downstream since it is a realistic date), it is not coalesced with the other parsed columns and is reported separately in the output.
}
\examples{
library(tidyverse)
test_data <- tibble(A = c("91884", "", NA, "091884", "09984", "090984", "09/18/1984", "1984-09-18", "9/18/84", "9/9/84", "9-9-84", "9-18-84", "12-18-84"))


test_data \%>\%
        parse_date_col(A)
}
\seealso{
\code{\link[dplyr]{mutate}}
 \code{\link[lubridate]{ymd}}
}
