% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_columns_with_na_rm.R
\name{paste_columns_with_na_rm}
\alias{paste_columns_with_na_rm}
\title{Pastes given columns by row while removing true NAs}
\usage{
paste_columns_with_na_rm(.data, new_col_name, ..., collapse = " ",
  sep = "")
}
\arguments{
\item{.data}{dataframe}

\item{new_col_name}{name of the new column}

\item{...}{columns to be pasted together}
}
\description{
Pastes given columns by row while removing true NAs
}
