% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_count.R
\name{pivot_count}
\alias{pivot_count}
\title{Get count and distinct count of a column}
\usage{
pivot_count(dataframe, col)
}
\value{
A dataframe with columns "Variable" for input column name, "Desc" for description of the count type, and "Count" for the value.
}
\description{
Get count and distinct count of a column
}
