% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize.R
\name{randomize}
\alias{randomize}
\title{Randomize a Dataframe}
\usage{
randomize(data)
}
\arguments{
\item{data}{A dataframe or tibble.}
}
\description{
Randomize a Dataframe
}
\examples{
library(tidyverse)
test_data <- 
        tibble(
                Group = sample(c("Apple", "Pear"), size = 10, replace = TRUE),
                A     = sample(c(NA_integer_, 1:3), size = 10, replace = TRUE),
                B     = sample(c(NA_integer_, 4:6), size = 10, replace = TRUE)
        )


randomize(test_data)

}
