% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{release}
\alias{release}
\title{Release all the columns in a dataframe into vectors of the same name}
\usage{
release(dataframe)
}
\description{
Release all the columns in a dataframe into vectors of the same name
}
\seealso{
\code{\link[dplyr]{select}}
 \code{\link[purrr]{map2}}
}
