% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{release_df}
\alias{release_df}
\title{Release all the columns in a dataframe into vectors of the same name}
\usage{
release_df(dataframe)
}
\description{
Release all the columns in a dataframe into vectors of the same name
}
