% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_all_replace.R
\name{rename_all_replace}
\alias{rename_all_replace}
\title{Replace a pattern in all column names}
\usage{
rename_all_replace(dataframe, pattern, replacement)
}
\description{
Replace a pattern in all column names
}
