% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_all_with_prefix.R
\name{rename_all_with_prefix}
\alias{rename_all_with_prefix}
\title{Rename all columns with a prefix}
\usage{
rename_all_with_prefix(dataframe, prefix)
}
\description{
Rename all columns with a prefix
}
