% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename_all_with_replace}
\alias{rename_all_with_replace}
\title{Replace a pattern in all column names}
\usage{
rename_all_with_replace(data, pattern, replacement)
}
\description{
Replace a pattern in all column names
}
