% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename_as}
\alias{rename_as}
\title{FUNCTION_TITLE}
\usage{
rename_as(data, new_colnames)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{.x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[dplyr]{select_all}},\code{\link[dplyr]{funs}}
}
