% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_at_.R
\name{rename_at_replace}
\alias{rename_at_replace}
\title{Replace a pattern in all column names}
\usage{
rename_at_replace(dataframe, ..., pattern, replacement)
}
\description{
Replace a pattern in all column names
}
