% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_fields.R
\name{rename_fields}
\alias{rename_fields}
\title{Rename Fields with a Character Vector}
\usage{
rename_fields(data, new_colnames)
}
\arguments{
\item{new_colnames}{Character vector of new names with same length as ncols of data argument.}
}
\description{
Rename the fields of a dataframe with 
a vector of new field names.
}
\seealso{
\code{\link[dplyr]{rename_at}}
}
