% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multibyte_chars.R
\name{rm_multibyte_chars}
\alias{rm_multibyte_chars}
\title{Remove all multibyte characters in the dataframe}
\usage{
rm_multibyte_chars(data)
}
\description{
Remove all multibyte characters in the dataframe
}
