% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_row_by_number.R
\name{rm_row_by_number}
\alias{rm_row_by_number}
\title{Removes a row by its number}
\usage{
rm_row_by_number(dataframe, row.number)
}
\arguments{
\item{row.number}{integer}
}
\description{
Removes a row by its number
}
