% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated2.R
\name{sample_n_or_bust}
\alias{sample_n_or_bust}
\title{Sample a dataframe for n obs and returns full dataframe if n > nrow}
\usage{
sample_n_or_bust(dataframe, n, ...)
}
\arguments{
\item{n}{number of observations}

\item{...}{additional arguments for dplyr::sample_n function}
}
\description{
Sample a dataframe for n obs and returns full dataframe if n > nrow
}
