% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_.R
\name{select_distinct}
\alias{select_distinct}
\title{Selects columns and get the distinct values back}
\usage{
select_distinct(data, ...)
}
\description{
Selects columns and get the distinct values back
}
\seealso{
\code{\link[dplyr]{select}},\code{\link[dplyr]{distinct}}
}
