% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_distinct_column.R
\name{select_distinct_column}
\alias{select_distinct_column}
\title{Get all distinct values for a particular column}
\usage{
select_distinct_column(dataframe, col)
}
\description{
Get all distinct values for a particular column
}
