% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_.R
\name{select_if_all_na}
\alias{select_if_all_na}
\title{Select columns that are all NA}
\usage{
select_if_all_na(dataframe)
}
\description{
Select columns that are all NA
}
\seealso{
\code{\link[dplyr]{select_all}}
}
