% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_select.R
\name{special_select}
\alias{special_select}
\title{Selects columns called upon without prefix that matches dataframe name}
\usage{
special_select(dataframe, ...)
}
\arguments{
\item{dataframe}{dataframe where the R object it belongs to has the same name as the prefix of the columns}
}
\description{
Selects columns called upon without prefix that matches dataframe name
}
