% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split_by}
\alias{split_by}
\title{Split a dataframe into a list with the column removed}
\usage{
split_by(data, column, drop = FALSE, sep = ".", lex.order = FALSE)
}
\description{
Split a dataframe into a list with the column removed
}
