% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_.R
\name{summarize_cardinality}
\alias{summarize_cardinality}
\title{Summarize One to Many Relationships}
\usage{
summarize_cardinality(data, source_cols, target_cols)
}
\description{
A wrapper around the `dplyr::count()` function that 
includes the source values, which are eliminated 
in using that function alone.
}
\seealso{
\code{\link[dplyr]{tidyeval-compat}},\code{\link[dplyr]{mutate-joins}},\code{\link[dplyr]{select_all}},\code{\link[dplyr]{distinct}},\code{\link[dplyr]{group_by_all}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{group_by}}
}
