% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated2.R
\name{summarize_grouped_n}
\alias{summarize_grouped_n}
\title{Summarize groups by count}
\usage{
summarize_grouped_n(dataframe, ..., desc = FALSE)
}
\description{
Summarize groups by count
}
\seealso{
\code{\link[dplyr]{tidyeval-compat}},\code{\link[dplyr]{group_by}},\code{\link[dplyr]{summarise}},\code{\link[dplyr]{arrange}},\code{\link[dplyr]{desc}}
}
\keyword{internal}
