% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_.R
\name{summarize_max}
\alias{summarize_max}
\title{Summarizes each column with max value}
\usage{
summarize_max(data, ..., na.rm = TRUE, grouper)
}
\arguments{
\item{data}{A dataframe or tibble.}

\item{...}{column names for vectors of any data class}

\item{na.rm}{TRUE if true NA are to be removed.}

\item{grouper}{(optional) Group by column.}
}
\description{
Summarizes each column with max value
}
\examples{
library(tidyverse)
test_data <- 
        tibble(
                Group = sample(c("Apple", "Pear"), size = 10, replace = TRUE),
                A     = sample(c(NA_integer_, 1:3), size = 10, replace = TRUE),
                B     = sample(c(NA_integer_, 4:6), size = 10, replace = TRUE),
                C     = sample(c(NA_real_, seq(from = 6.01, to = 6.09, by = 0.01)), size = 10, replace = TRUE)
        )


# Checking that all numeric cols are captured
all_numeric_cols(data = test_data)

# Ungrouped Summary
summarize_max(test_data)
summarize_max(test_data, A)

# Grouped Summary
summarize_max(test_data,
              grouper = Group)

summarize_max(test_data,
              B,
              na.rm = TRUE,
              grouper = Group)

summarize_max(test_data,
              B,
              na.rm = FALSE,
              grouper = Group)
}
\seealso{
\code{\link[dplyr]{tidyeval-compat}}

Other summarize functions: 
\code{\link{value_count}()}
}
\concept{summarize functions}
