% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_max_value_per_column.R
\name{summarize_max_value_per_column}
\alias{summarize_max_value_per_column}
\title{Summarizes each column with max value}
\usage{
summarize_max_value_per_column(.data, ..., na.rm = TRUE)
}
\arguments{
\item{...}{column names for vectors of any data class}

\item{na.rm}{TRUE if true NA are to be removed. Default is TRUE.}
}
\description{
Summarizes each column with max value
}
