% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_.R
\name{summarize_numeric}
\alias{summarize_numeric}
\title{Summarize Numeric Columns with Standard Summary Functions}
\usage{
summarize_numeric(data, ..., grouper)
}
\arguments{
\item{data}{A dataframe or tibble.}

\item{...}{column names for vectors of any data class}

\item{grouper}{(optional) Group by column.}

\item{na.rm}{TRUE if true NA are to be removed.}
}
\description{
Summarize Numeric Columns with Standard Summary Functions
}
\examples{
library(tidyverse)
test_data <- 
        tibble(
                Group = sample(c("Apple", "Pear"), size = 10, replace = TRUE),
                A     = sample(c(NA_integer_, 1:3), size = 10, replace = TRUE),
                B     = sample(c(NA_integer_, 4:6), size = 10, replace = TRUE),
                C     = sample(c(NA_real_, seq(from = 6.01, to = 6.09, by = 0.01)), size = 10, replace = TRUE)
        )


summarize_numeric(test_data)
summarize_numeric(test_data, grouper = Group)
summarize_numeric(test_data, A,grouper = Group)
summarize_numeric(test_data, B,grouper = Group)
}
\seealso{
\code{\link[dplyr]{tidyeval-compat}},\code{\link[dplyr]{select}},\code{\link[dplyr]{summarise_all}},\code{\link[dplyr]{vars}},\code{\link[dplyr]{reexports}},\code{\link[dplyr]{group_by_all}},\code{\link[dplyr]{group_by}}
}
