% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_numerics.R
\name{summarize_numerics}
\alias{summarize_numerics}
\title{Summarize a variable by the standard summary functions}
\usage{
summarize_numerics(.data, ...)
}
\arguments{
\item{...}{grouping vars. If missing, all variables will be summarized.}
}
\description{
Summarize a variable by the standard summary functions
}
