% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_.R
\name{row_count}
\alias{row_count}
\title{Get the number of times a row appears in a dataframe}
\usage{
row_count(data, desc = TRUE)
}
\description{
Shortcut for `group_by_all()` followed by `count()`.
 \code{\link[cave]{vector_to_string}}
 \code{\link[dplyr]{tidyeval-compat}},\code{\link[dplyr]{group_by}},\code{\link[dplyr]{summarise}},\code{\link[dplyr]{arrange}},\code{\link[dplyr]{desc}},\code{\link[dplyr]{group_by_all}}
}
