% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_values.R
\name{summarize_values}
\alias{summarize_values}
\title{Summarize the counts of values in columns}
\usage{
summarize_values(dataframe, ..., desc = TRUE)
}
\arguments{
\item{...}{grouping vars. If missing, groups and summarizes based on all columns.}
}
\description{
Summarize the counts of values in columns
}
\examples{
data("mtcars")
summarize_values(mtcars)
summarize_values(mtcars, mpg, cyl)
}
