% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated2.R
\name{summarize_var_group}
\alias{summarize_var_group}
\title{Summarize a variable by the standard summary functions}
\usage{
summarize_var_group(.data, ...)
}
\arguments{
\item{...}{grouping vars. If missing, all variables will be summarized.}
}
\description{
Summarize a variable by the standard summary functions
}
\seealso{
\code{\link[dplyr]{tidyeval-compat}},\code{\link[dplyr]{group_by_all}},\code{\link[dplyr]{summarise_all}},\code{\link[dplyr]{mutate_all}}
 \code{\link[tidyr]{pivot_longer}},\code{\link[tidyr]{pivot_wider}}
}
\keyword{internal}
