% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_variable_metrics.R
\name{summarize_variable_metrics}
\alias{summarize_variable_metrics}
\title{Summarize a variable by distinct counts of values it has}
\usage{
summarize_variable_metrics(dataframe, ...)
}
\arguments{
\item{...}{grouping vars. If missing, all variables will be counted}
}
\description{
Summarize a variable by distinct counts of values it has
}
