% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_variable_values.R
\name{summarize_variable_values}
\alias{summarize_variable_values}
\title{Summarize a variable by distinct counts of values it has}
\usage{
summarize_variable_values(dataframe, ..., as_expr = FALSE, collapse = "|")
}
\arguments{
\item{...}{grouping vars. If missing, all variables will be counted}
}
\description{
Summarize a variable by distinct counts of values it has
}
