% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_.R
\name{summarize_variables}
\alias{summarize_variables}
\title{Summarize a Variable}
\usage{
summarize_variables(
  data,
  incl_num_calc = TRUE,
  names_to = "Variable",
  values_to = "Value",
  grouper
)
}
\arguments{
\item{data}{A dataframe or tibble.}

\item{incl_num_calc}{If TRUE, includes an additional dataframe of summary statistics on the numeric columns in the dataframe.}

\item{names_to}{Passed to `tidyr::pivot_longer()`.}

\item{values_to}{Passed to `tidyr::pivot_longer()`.}

\item{grouper}{(Optional) Group by column.}
}
\description{
Summarize a Variable
}
\examples{
library(tidyverse)
test_data <- 
        tibble(
                Group = sample(c("Apple", "Pear"), size = 10, replace = TRUE),
                A     = sample(c(NA_integer_, 1:3), size = 10, replace = TRUE),
                B     = sample(c(NA_integer_, 4:6), size = 10, replace = TRUE),
                C     = sample(c(NA_real_, seq(from = 6.01, to = 6.09, by = 0.01)), size = 10, replace = TRUE)
        )


summarize_variables(data = test_data,
                    incl_num_calc = FALSE)

summarize_variables(data = test_data,
                    incl_num_calc = TRUE)

summarize_variables(data = test_data,
                    incl_num_calc = TRUE,
                    grouper = Group)

summarize_variables(data = test_data,
                    incl_num_calc = FALSE,
                    grouper = Group)


observation_count(data = test_data)


value_count(data = test_data)
value_count(data = test_data,
            grouper = Group)
}
\seealso{
\code{\link[tidyr]{pivot_longer}}
 \code{\link[dplyr]{reexports}},\code{\link[dplyr]{group_by_all}},\code{\link[dplyr]{vars}},\code{\link[dplyr]{summarise_all}},\code{\link[dplyr]{select}}

Other summary functions: 
\code{\link{observation_count}()}
}
\concept{summary functions}
