% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_grouped_vars.R
\name{summarize_vars_grouped}
\alias{summarize_vars_grouped}
\title{Summarize a variable by the standard summary functions}
\usage{
summarize_vars_grouped(.data, ...)
}
\arguments{
\item{...}{grouping vars. If missing, all variables will be summarized.}
}
\description{
Summarize a variable by the standard summary functions
}
