% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_unique_values.R
\name{survey_unique_values}
\alias{survey_unique_values}
\title{Get all the unique values in each column printed out as a list of vectors}
\usage{
survey_unique_values(dataframe, alphabetical_order = TRUE)
}
\arguments{
\item{dataframe}{dataframe with target columns to iterate over}

\item{alphabetical_order}{defaults to TRUE where columns will be returned in alphabetical order.}
}
\description{
Get all the unique values in each column printed out as a list of vectors
}
