% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_to_df.R
\name{turn_vector_into_dataframe}
\alias{turn_vector_into_dataframe}
\title{Converts vector to a dataframe with names as column names}
\usage{
turn_vector_into_dataframe(vector)
}
\arguments{
\item{named_vector}{vector of any length with names}
}
\description{
Converts vector to a dataframe with names as column names
}
