% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmask_column_prefixes.R
\name{unmask_column_prefixes}
\alias{unmask_column_prefixes}
\title{For R objects where column names are prefixed for object names, restores the full column names}
\usage{
unmask_column_prefixes(dataframe)
}
\description{
For R objects where column names are prefixed for object names, restores the full column names
}
