% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_.R
\name{value_count}
\alias{value_count}
\title{Value Count for a Dataframe}
\usage{
value_count(data, names_to = "Variable", values_to = "Value", desc = TRUE)
}
\arguments{
\item{data}{A dataframe or tibble.}

\item{names_to}{passed to tidyr pivot_longer()}

\item{values_to}{passed to tidyr pivot_longer()}
}
\description{
Value Count for a Dataframe
}
\seealso{
\code{\link[tidyr]{pivot_longer}}
 \code{\link[dplyr]{reexports}},\code{\link[dplyr]{count}},\code{\link[dplyr]{tidyeval-compat}},\code{\link[dplyr]{arrange}}

Other summarize functions: 
\code{\link{summarize_max}()}
}
\concept{summarize functions}
