% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_as_dataframe.R
\name{vector_as_dataframe}
\alias{vector_as_dataframe}
\title{Convert a vector to a dataframe
This function takes a vector and returns a dataframe with one row of values provided by the vector. Column names are those of the vector or the generic names in cases of unnamed vectors.}
\usage{
vector_as_dataframe(vector)
}
\arguments{
\item{vector}{vector of any length with names}
}
\value{
dataframe
}
\description{
Convert a vector to a dataframe
This function takes a vector and returns a dataframe with one row of values provided by the vector. Column names are those of the vector or the generic names in cases of unnamed vectors.
}
\examples{
 x <- c("a", "b", "c")
 vector_as_dataframe(x) 
}
