% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_to_tibble.R
\name{vector_to_tibble}
\alias{vector_to_tibble}
\title{Take a single vector of values and convert it to a dataframe}
\usage{
vector_to_tibble(vector, new_col)
}
\description{
This function was written in cases were a list of vectors needed to be aggregated either by bind_cols and bind_rows of the dplyr package.
}
