% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_args.R
\name{wrapper_args}
\alias{wrapper_args}
\title{Tidyverse Wrapper Arguments}
\arguments{
\item{data}{A dataframe or tibble.}

\item{col}{Target column.}
}
\description{
Tidyverse Wrapper Arguments
}
\keyword{internal}
