% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_as_dataframe.R
\docType{data}
\name{x}
\alias{x}
\title{Convert a vector to a dataframe
This function takes a vector and}
\format{
An object of class \code{character} of length 3.
}
\usage{
x
}
\arguments{
\item{vector}{vector of any length with names}
}
\value{
dataframe
}
\description{
Convert a vector to a dataframe
This function takes a vector and
}
\keyword{datasets}
