% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_as_integer.R
\name{arrange_as_integer}
\alias{arrange_as_integer}
\title{Arrange a column any data class as an integer}
\usage{
arrange_as_integer(.data, column, desc = FALSE)
}
\description{
Arrange a column any data class as an integer
}
