% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_as_integer.R
\name{arrange_as_integer}
\alias{arrange_as_integer}
\title{Arrange by Column as an Integer}
\usage{
arrange_as_integer(.data, column, desc = FALSE)
}
\description{
Perform an arrange function call on a dataframe with the values of the target column as an integer class.
}
