% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated2.R
\name{arrange_as_integer}
\alias{arrange_as_integer}
\title{Arrange by a Given Column as Integer}
\usage{
arrange_as_integer(.data, column, desc = FALSE)

arrange_as_integer(.data, column, desc = FALSE)
}
\arguments{
\item{.data}{PARAM_DESCRIPTION}

\item{column}{PARAM_DESCRIPTION}

\item{desc}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
Perform an arrange function call on a dataframe with the values of the target column as an integer class.

(Deprecated) Perform an arrange function call on a dataframe with the values of the target column as an integer class.
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{tidyeval-compat}},\code{\link[dplyr]{arrange}}

\code{\link[dplyr]{tidyeval-compat}},\code{\link[dplyr]{arrange}}
}
