% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_by_nchar.R
\name{arrange_by_nchar}
\alias{arrange_by_nchar}
\title{Arrange dataframe based on nchar of a given column}
\usage{
arrange_by_nchar(dataframe, nchar_col, desc = FALSE)
}
\description{
Arrange dataframe based on nchar of a given column
}
