% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blank_tibble.R
\name{blank_tibble}
\alias{blank_tibble}
\title{Create a empty dataframe with column names
This function takes a character vector and creates a 0-row dataframe with the vector as column name.}
\usage{
blank_tibble(column_names)
}
\arguments{
\item{column_names}{vector of column names}
}
\description{
Create a empty dataframe with column names
This function takes a character vector and creates a 0-row dataframe with the vector as column name.
}
