% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blank_tibble.R
\name{blank_tibble}
\alias{blank_tibble}
\title{Takes a character vector and creates a 0-row dataframe with the vectors as column names}
\usage{
blank_tibble(column_names)
}
\description{
Takes a character vector and creates a 0-row dataframe with the vectors as column names
}
