% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_mr_clean.R
\name{call_mr_clean}
\alias{call_mr_clean}
\title{General Dataframe Cleanup
Convert all columns to character class and trim all left and right white spaces.}
\usage{
call_mr_clean(dataframe)
}
\arguments{
\item{dataframe}{input dataframe}
}
\description{
General Dataframe Cleanup
Convert all columns to character class and trim all left and right white spaces.
}
