% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_list.R
\name{cols_to_list}
\alias{cols_to_list}
\title{Dataframe to a List of Vectors}
\usage{
cols_to_list(data, ...)
}
\arguments{
\item{data}{A dataframe or tibble.}

\item{...}{(optional) Columns to select for.}
}
\description{
Convert a dataframe into a list of vectors for each column. If ... is empty, the entire dataframe will be broken down.
}
\seealso{
\code{\link[dplyr]{select}}

Other dataframe <> list functions: 
\code{\link{list_to_tibble}()}
}
\concept{dataframe <> list functions}
