% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_to_list.R
\name{cols_to_list}
\alias{cols_to_list}
\title{Select cols and convert them to a list}
\usage{
cols_to_list(data, ...)
}
\arguments{
\item{id_column_name}{optional string of the column designated as the identifier that should remain in the output.}
}
\description{
This function takes a dataframe and breaks it down into a list by column. The columns can be explicitly provided or if ... is empty, the entire dataframe will be broken down.
}
\seealso{
\code{\link[dplyr]{tidyeval-compat}},\code{\link[dplyr]{select}}
 \code{\link[purrr]{map}}
}
