% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_all.R
\name{filter_all_grepl_any}
\alias{filter_all_grepl_any}
\title{Filter at a variable that contains or doesn't contain a phrase}
\usage{
filter_all_grepl_any(
  data,
  grepl_phrase,
  evaluates_to = TRUE,
  ignore.case = TRUE
)
}
\arguments{
\item{grepl_phrase}{phrase that is being filtered for}

\item{evaluates_to}{whether the filter is for a grepl evaluation of TRUE or FALSE. Defaults to TRUE.}

\item{ignore.case}{boolean the ignore.case argument of grepl function}

\item{dataframe}{input dataframe}

\item{col}{column to filter}
}
\description{
Filter at a variable that contains or doesn't contain a phrase
}
\seealso{
\code{\link[dplyr]{filter_all}},\code{\link[dplyr]{vars}},\code{\link[dplyr]{all_vars}}
}
