% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_words_in_str.R
\name{filter_by_words_in_str}
\alias{filter_by_words_in_str}
\title{Loop over searching for words in a string in a dataframe}
\usage{
filter_by_words_in_str(string, split, dataframe, col)
}
\description{
Loop over searching for words in a string in a dataframe
}
