% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_words_in_str.R
\name{filter_by_words_in_str}
\alias{filter_by_words_in_str}
\title{Loop a filter over words in a string
This function splits a string based on the provide split argument, and recursively filters a dataframe for those words at a given column.}
\usage{
filter_by_words_in_str(string, split, dataframe, col)
}
\arguments{
\item{string}{string of words or other tokens to loop the filter over}

\item{split}{split argument for the strsplit function applied to the string parameter}

\item{dataframe}{input data}

\item{col}{column to filter against}
}
\description{
Loop a filter over words in a string
This function splits a string based on the provide split argument, and recursively filters a dataframe for those words at a given column.
}
