% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_most_recent_obs_by_group.R
\name{filter_most_recent_obs_by_group}
\alias{filter_most_recent_obs_by_group}
\title{Filters most recent values in dataframe}
\usage{
filter_most_recent_obs_by_group(dataframe, group_by_col, index_date_col)
}
\arguments{
\item{group_by_col}{grouping variable}

\item{index_date_col}{variable that will be used to filter by time}
}
\description{
Filters most recent values in dataframe
}
