% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-na.R
\name{grep_empty}
\alias{grep_empty}
\title{FUNCTION_TITLE}
\usage{
grep_empty(vector)
}
\arguments{
\item{vector}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
