% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/summary_fun_list.R
\name{loadSummaryFnLibrary}
\alias{loadSummaryFnLibrary}
\alias{summary_fun_list}
\title{Load Summary Function Library}
\usage{
loadSummaryFnLibrary()

summary_fun_list()
}
\description{
Load Summary Function Library

Load Summary Function Library
}
\seealso{
\code{\link[centipede]{no_na}}

\code{\link[centipede]{no_na}}
}
