% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_rowsum.R
\name{mutate_rowsum}
\alias{mutate_rowsum}
\title{Get the sum of the rows base on the designated columns}
\usage{
mutate_rowsum(dataframe, ...)
}
\description{
Get the sum of the rows base on the designated columns
}
