% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_all_prefix.R
\name{rename_all_prefix}
\alias{rename_all_prefix}
\title{Rename all columns with a prefix}
\usage{
rename_all_prefix(dataframe, prefix)
}
\description{
Rename all columns with a prefix
}
\seealso{
\code{\link[dplyr]{select_all}}
}
