% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample_all_or_n}
\alias{sample_all_or_n}
\title{Sample a dataframe for n obs and returns full dataframe if n > nrow}
\usage{
sample_all_or_n(data, n, replace = TRUE, ...)
}
\arguments{
\item{n}{number of observations}

\item{...}{additional arguments for dplyr::sample_n function}
}
\description{
Sample a dataframe for n obs and returns full dataframe if n > nrow
}
