% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_distinct.R
\name{select_distinct}
\alias{select_distinct}
\title{Selects columns and get the distinct values back}
\usage{
select_distinct(dataframe, ...)
}
\description{
Selects columns and get the distinct values back
}
