% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copysplit_deselect.R, R/split_deselect.R
\name{split_deselect}
\alias{split_deselect}
\title{Split a dataframe into a list with the column removed}
\usage{
split_deselect(x, column_name, ...)

split_deselect(x, column_name, ...)
}
\description{
Split a dataframe into a list with the column removed

Split a dataframe into a list with the column removed
}
