% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_max.R
\name{summarize_max}
\alias{summarize_max}
\title{Summarizes each column with max value}
\usage{
summarize_max(.data, ..., na.rm = TRUE)
}
\arguments{
\item{...}{column names for vectors of any data class}

\item{na.rm}{TRUE if true NA are to be removed. Default is TRUE.}
}
\description{
Summarizes each column with max value
}
\seealso{
\code{\link[dplyr]{tidyeval-compat}}
}
