% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_numeric_vars.R
\name{summarize_numeric_vars}
\alias{summarize_numeric_vars}
\title{Summarize a variable by the standard summary functions}
\usage{
summarize_numeric_vars(.data, ...)
}
\arguments{
\item{...}{grouping vars. If missing, all variables will be summarized.}
}
\description{
Summarize a variable by the standard summary functions
}
