% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_rows.R
\name{summarize_rows}
\alias{summarize_rows}
\title{Summarize the counts of values in columns}
\usage{
summarize_rows(dataframe, ..., desc = TRUE)
}
\arguments{
\item{...}{grouping vars. If missing, groups and summarizes based on all columns.}
}
\description{
To group by more than 1 variable can be summarized using summarize_grouped_vars().
}
