% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_values.R
\name{summarize_values}
\alias{summarize_values}
\title{Summarize the counts of values in columns}
\usage{
summarize_values(.data, ..., desc = TRUE)
}
\arguments{
\item{...}{grouping vars. If missing, groups and summarizes based on all columns.}
}
\description{
To group by more than 1 variable can be summarized using summarize_grouped_vars().
}
\seealso{
\code{\link[rlang]{as_name}}
 \code{\link[purrr]{map}},\code{\link[purrr]{set_names}}
 \code{\link[dplyr]{select}},\code{\link[dplyr]{select_all}},\code{\link[dplyr]{group_by_all}},\code{\link[dplyr]{bind}},\code{\link[dplyr]{summarise}}
 \code{\link[stringr]{str_replace}}
}
