% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated2.R
\name{summarize_values}
\alias{summarize_values}
\title{Summarize the counts of values in columns}
\usage{
summarize_values(.data, ..., names_to = "name", values_to = "value")
}
\arguments{
\item{.data}{data frame}

\item{...}{group by variables}

\item{names_to}{passed to tidyr pivot_longer()}

\item{values_to}{passed to tidyr pivot_longer()}
}
\description{
(Deprecated) To group by more than 1 variable can be summarized using summarize_grouped_vars().
}
\seealso{
\code{\link[rlang]{as_name}}
 \code{\link[purrr]{map}},\code{\link[purrr]{set_names}}
 \code{\link[dplyr]{select}},\code{\link[dplyr]{select_all}},\code{\link[dplyr]{group_by_all}},\code{\link[dplyr]{bind}},\code{\link[dplyr]{summarise}}
 \code{\link[stringr]{str_replace}}

\code{\link[rlang]{as_name}}
 \code{\link[dplyr]{mutate_all}},\code{\link[dplyr]{group_by_all}},\code{\link[dplyr]{summarise_all}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{rename}},\code{\link[dplyr]{select}},\code{\link[dplyr]{reexports}}
 \code{\link[tidyr]{pivot_longer}}
}
