% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_rows.R, R/summarize_values.R
\name{summarize_values}
\alias{summarize_values}
\title{Summarize the counts of values in columns}
\usage{
summarize_values(.data, ..., desc = TRUE)

summarize_values(.data, ..., desc = TRUE)
}
\arguments{
\item{...}{grouping vars. If missing, groups and summarizes based on all columns.}
}
\description{
To group by more than 1 variable can be summarized using summarize_grouped_vars().

To group by more than 1 variable can be summarized using summarize_grouped_vars().
}
