% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_.R
\name{summarize_variables}
\alias{summarize_variables}
\title{Summarize a Variable}
\usage{
summarize_variables(
  data,
  incl_num_calc = TRUE,
  names_to = "Variable",
  values_to = "Value"
)
}
\arguments{
\item{incl_num_calc}{If TRUE, includes an additional dataframe of summary statistics on the numeric columns in the dataframe.}
}
\description{
Summarize a Variable
}
\examples{
library(tidyverse)
test_data <- 
        tibble(
                Group = sample(c("Apple", "Pear"), size = 10, replace = TRUE),
                A     = sample(c(NA_integer_, 1:3), size = 10, replace = TRUE),
                B     = sample(c(NA_integer_, 4:6), size = 10, replace = TRUE),
                C     = sample(c(NA_real_, seq(from = 6.01, to = 6.09, by = 0.01)), size = 10, replace = TRUE)
        )


summarize_variables(data = test_data,
                    incl_num_calc = FALSE)

summarize_variables(data = test_data,
                    incl_num_calc = TRUE)
}
\seealso{
\code{\link[tidyr]{pivot_longer}}
 \code{\link[dplyr]{reexports}},\code{\link[dplyr]{group_by_all}},\code{\link[dplyr]{vars}},\code{\link[dplyr]{summarise_all}},\code{\link[dplyr]{select}}
}
\concept{summary functions}
