% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_variables.R
\name{summarize_variables}
\alias{summarize_variables}
\title{Summarize a Variable}
\usage{
summarize_variables(.data, ..., incl_expr = FALSE)
}
\arguments{
\item{...}{grouping vars. If missing, all variables will be summarized.}
}
\description{
Summarize a Variable
}
\seealso{
\code{\link[cave]{vector_to_string}}
 \code{\link[dplyr]{tidyeval-compat}},\code{\link[dplyr]{summarise_all}},\code{\link[dplyr]{mutate_all}}
 \code{\link[tidyr]{pivot_longer}},\code{\link[tidyr]{pivot_wider}}
}
