% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_as_dataframe.R
\name{vector_as_dataframe}
\alias{vector_as_dataframe}
\title{Converts vector to a dataframe with names as column names}
\usage{
vector_as_dataframe(vector)
}
\arguments{
\item{vector}{vector of any length with names}
}
\value{
dataframe
}
\description{
Converts vector to a dataframe with names as column names
}
\examples{
x <- c("a", "b", "c")
vector_as_dataframe(x) 
}
