% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated2.R
\name{vector_to_tibble}
\alias{vector_to_tibble}
\title{Convert a Vector to a Tibble}
\usage{
vector_to_tibble(vector, new_col)
}
\arguments{
\item{vector}{Vector that will become the column in the tibble}

\item{new_col}{Name of the new column}
}
\value{
A tibble with 1 column with `new_col` as the name.
}
\description{
Convert a Vector to a Tibble
}
\seealso{
\code{\link[dplyr]{tidyeval-compat}}
 \code{\link[tibble]{tibble}}
}
